Make Question state controlled
