+++ quiz-us-components/src/components/RichTextEditor/index.js
@@ -295,6 +295,7 @@ class RichTextEditor extends React.Component {
 
   onChange = ({ value }) => {
     this.setState({ value });
+    this.props.updateParentState(value);
   };
 
   /**
